Update allocations to include test_usd for everyone as well as a less obscene amount of allocations to osiris (#2552)

+++ app/src/governance/view.rs
@@ -261,8 +261,7 @@ pub trait StateReadExt: StateRead + crate::stake::StateReadExt {
             };
 
         // Check that the unbonded amount is correct relative to that exchange rate
-        if rate_data.unbonded_amount(value.amount.value() as u64) != unbonded_amount.value() as u64
-        {
+        if rate_data.unbonded_amount(value.amount.value()) != unbonded_amount.value() {
             anyhow::bail!(
                 "unbonded amount {}{} does not correspond to {} staked delegation tokens for validator {} using the exchange rate at the start of proposal {}",
                 unbonded_amount,
